Re-organized functions into dedicated folders
