+++ dbtwiz/commands/__init__.py
@@ -3,13 +3,13 @@ from typing import Annotated
 
 import typer
 
-from dbtwiz.manifest import Manifest
-from dbtwiz.target import Target
-from dbtwiz.utils.decorators import examples
+from dbtwiz.config.user import UserConfig
+from dbtwiz.dbt.manifest import Manifest
+from dbtwiz.dbt.target import Target
+from dbtwiz.helpers.decorators import examples
 
 from .backfill import backfill as command_backfill
 from .build import build as command_build
-from .config import config as command_config
 from .test import test as command_test
 
 app = typer.Typer()
@@ -281,5 +281,10 @@ def config(
     setting: Annotated[str, typer.Argument(help="Configuration setting")],
     value: Annotated[str, typer.Argument(help="Configuration value")],
 ):
-    """Update configuration setting"""
-    command_config(setting, value)
+    """Update user configuration setting"""
+    if ":" in setting:
+        section, key = setting.split(":")
+    else:
+        section, key = "general", setting
+
+    UserConfig().update(section, key, value)
