+++ dbtwiz/commands/backfill.py
@@ -6,9 +6,9 @@ from textwrap import dedent
 
 from jinja2 import Template
 
-from dbtwiz.auth import ensure_auth
-from dbtwiz.config import project_config, project_dbtwiz_path
-from dbtwiz.logging import debug, info
+from dbtwiz.config.project import project_config, project_dbtwiz_path
+from dbtwiz.gcp.auth import ensure_auth
+from dbtwiz.helpers.log_types import debug, info
 
 MAX_CONCURRENT_TASKS = 8
 YAML_FILE = project_dbtwiz_path("backfill-cloudrun.yaml")
