+++ dbtwiz/helpers/typer_docs.py
@@ -6,7 +6,7 @@ import typer
 from typer.models import ArgumentInfo, DefaultPlaceholder, OptionInfo
 
 
-def extract_param_info(param) -> dict:
+def _extract_param_info(param) -> dict:
     """Extract all parameter information including both long and short flags"""
     param_info = {
         "name": param.name,
@@ -59,7 +59,7 @@ def extract_param_info(param) -> dict:
     return param_info
 
 
-def get_decorator_docs(command_func, decorator_name, docs_header) -> str:
+def _get_decorator_docs(command_func, decorator_name, docs_header) -> str:
     decorator_value = getattr(command_func, decorator_name, [])
 
     # Generate docs section
@@ -74,7 +74,7 @@ def get_decorator_docs(command_func, decorator_name, docs_header) -> str:
     return section_markdown
 
 
-def generate_markdown(
+def _generate_markdown(
     app_name: str,
     full_command_path: List[str],
     command_func,
@@ -110,7 +110,7 @@ def generate_markdown(
         if name == "return":
             continue
 
-        param_info = extract_param_info(param)
+        param_info = _extract_param_info(param)
 
         if param_info["is_option"] and param_info["flags"]:
             flags = [f"`{flag}`" for flag in param_info["flags"]]
@@ -120,14 +120,14 @@ def generate_markdown(
 
     markdown = f"# `{command_path_str}`\n\n{description}\n\n"
 
-    markdown += get_decorator_docs(command_func, "_command_description", None)
+    markdown += _get_decorator_docs(command_func, "_command_description", None)
 
     if required_args:
         markdown += "## Required arguments\n\n" + "\n".join(required_args) + "\n\n"
     if options:
         markdown += "## Options\n\n" + "\n".join(options)
 
-    markdown += get_decorator_docs(command_func, "_command_examples", "Examples")
+    markdown += _get_decorator_docs(command_func, "_command_examples", "Examples")
 
     # Check for changes before writing file
     if output_file.exists():
@@ -143,7 +143,7 @@ def generate_markdown(
         print(f"[x] Failed to write {output_file.relative_to(repo_root)}: {e}")
 
 
-def get_group_name(group) -> str:
+def _get_group_name(group) -> str:
     """Get the group name with fallbacks"""
     if hasattr(group, "name") and group.name:
         return group.name
@@ -152,19 +152,19 @@ def get_group_name(group) -> str:
     return "group"
 
 
-def should_have_docs(cmd) -> bool:
+def _should_have_docs(cmd) -> bool:
     """Determine if a command should have its own documentation file"""
     return hasattr(cmd, "callback") and cmd.callback
 
 
-def get_clean_command_name(cmd) -> str:
+def _get_clean_command_name(cmd) -> str:
     """Get a clean command name, handling DefaultPlaceholder and other cases"""
     if isinstance(cmd, DefaultPlaceholder):
         return str(cmd.value) if hasattr(cmd, "value") else "command"
     return str(cmd)
 
 
-def get_command_description(cmd) -> str:
+def _get_command_description(cmd) -> str:
     """Extract the command description from help or the callback info"""
     if hasattr(cmd, "help") and cmd.help:
         return cmd.help
@@ -175,13 +175,13 @@ def get_command_description(cmd) -> str:
     return "No description available."
 
 
-def process_command_group(group, current_path: List[str], level: int = 0) -> List[str]:
+def _process_command_group(group, current_path: List[str], level: int = 0) -> List[str]:
     """Process a command group and return a list of command entries"""
     command_list = []
 
     # Clean the current path
     cleaned_path = [
-        get_clean_command_name(cmd)
+        _get_clean_command_name(cmd)
         for cmd in current_path
         if not isinstance(cmd, DefaultPlaceholder)
     ]
@@ -189,11 +189,11 @@ def process_command_group(group, current_path: List[str], level: int = 0) -> Lis
     # Process commands first
     for cmd in getattr(group, "registered_commands", []):
         if hasattr(cmd, "callback") and cmd.callback:
-            cmd_name = get_clean_command_name(cmd.callback.__name__)
+            cmd_name = _get_clean_command_name(cmd.callback.__name__)
 
             full_path = cleaned_path + [cmd_name]
             safe_filename = "_".join(full_path).lower()
-            description = get_command_description(cmd)
+            description = _get_command_description(cmd)
             indent = "  " * level
             rel_docs_path = f"docs/{safe_filename}.md"
             command_list.append(
@@ -203,20 +203,20 @@ def process_command_group(group, current_path: List[str], level: int = 0) -> Lis
     # Then process groups (subcommands)
     for subgroup in getattr(group, "registered_groups", []):
         if hasattr(subgroup, "typer_instance"):
-            group_name = get_group_name(subgroup)
+            group_name = _get_group_name(subgroup)
             if group_name is None or isinstance(group_name, DefaultPlaceholder):
                 command_list.extend(
-                    process_command_group(subgroup.typer_instance, cleaned_path, level)
+                    _process_command_group(subgroup.typer_instance, cleaned_path, level)
                 )
                 continue
 
             full_path = cleaned_path + [group_name]
             indent = "  " * level
-            description = get_command_description(subgroup)
+            description = _get_command_description(subgroup)
             safe_filename = "_".join(full_path).lower()
             rel_docs_path = f"docs/{safe_filename}.md"
 
-            if should_have_docs(subgroup):
+            if _should_have_docs(subgroup):
                 command_list.append(
                     f"{indent}- [`{group_name}`]({rel_docs_path}) - {description}"
                 )
@@ -224,19 +224,19 @@ def process_command_group(group, current_path: List[str], level: int = 0) -> Lis
                 command_list.append(f"{indent}- `{group_name}` - {description}")
 
             command_list.extend(
-                process_command_group(subgroup.typer_instance, full_path, level + 1)
+                _process_command_group(subgroup.typer_instance, full_path, level + 1)
             )
 
     return command_list
 
 
-def generate_readme_command_list(app: typer.Typer, app_name: str) -> str:
+def _generate_readme_command_list(app: typer.Typer, app_name: str) -> str:
     """Generate a markdown list of all commands with links to their documentation"""
-    command_list = process_command_group(app, [])
+    command_list = _process_command_group(app, [])
     return "\n".join(command_list)
 
 
-def update_readme(app: typer.Typer, app_name: str, repo_root: Path, readme_path: Path):
+def _update_readme(app: typer.Typer, app_name: str, repo_root: Path, readme_path: Path):
     """Update README.md with the command list between the special comments"""
     if not readme_path.exists():
         print(f"[x] README.md not found at {readme_path.relative_to(repo_root)}")
@@ -246,20 +246,20 @@ def update_readme(app: typer.Typer, app_name: str, repo_root: Path, readme_path:
         with open(readme_path, "r", encoding="utf-8") as f:
             content = f.read()
 
-        start_marker = "[comment]: <> (START ACCESS CONFIG)"
-        end_marker = "[comment]: <> (END ACCESS CONFIG)"
+        start_marker = "[comment]: <> (START COMMAND DOCS)"
+        end_marker = "[comment]: <> (END COMMAND DOCS)"
 
         start_pos = content.find(start_marker)
         end_pos = content.find(end_marker)
 
         if start_pos == -1 or end_pos == -1:
-            print("[x] Could not find START/END ACCESS CONFIG markers in README.md")
+            print("[x] Couldn't find START/END COMMAND DOCS markers in README.md")
             return
 
         before_content = content[: start_pos + len(start_marker)]
         after_content = content[end_pos:]
 
-        command_list = generate_readme_command_list(app, app_name)
+        command_list = _generate_readme_command_list(app, app_name)
         new_content = f"{before_content}\n\n{command_list}\n\n{after_content}"
 
         # Check for changes before writing file
@@ -272,7 +272,7 @@ def update_readme(app: typer.Typer, app_name: str, repo_root: Path, readme_path:
         print(f"[x] Failed to update {readme_path.relative_to(repo_root)}: {e}")
 
 
-def process_commands(
+def _process_commands(
     app: typer.Typer,
     app_name: str,
     repo_root: Path,
@@ -286,7 +286,7 @@ def process_commands(
     for cmd in getattr(app, "registered_commands", []):
         if hasattr(cmd, "callback") and cmd.callback:
             cmd_name = cmd.callback.__name__
-            generate_markdown(
+            _generate_markdown(
                 app_name=app_name,
                 full_command_path=parent_commands + [cmd_name],
                 command_func=cmd.callback,
@@ -297,8 +297,8 @@ def process_commands(
     # Process groups
     for group in getattr(app, "registered_groups", []):
         if hasattr(group, "typer_instance"):
-            group_name = get_group_name(group)
-            process_commands(
+            group_name = _get_group_name(group)
+            _process_commands(
                 app=group.typer_instance,
                 app_name=app_name,
                 repo_root=repo_root,
@@ -307,12 +307,24 @@ def process_commands(
             )
 
 
-def update_docs(
+def generate(
     app: typer.Typer, app_name: str, repo_root: Path, docs_dir: Path, readme_path: Path
 ):
-    """Adds markdown files for each command and updates readme."""
-    process_commands(app=app, app_name=app_name, repo_root=repo_root, docs_dir=docs_dir)
-    update_readme(
+    """
+    Identifies commands in the given typer app, and for each command and subcommand,
+    creates markdown files documenting the command in the given docs_dir.
+
+    In addition, all the commands are added to `README.md` in the given readme path
+    between the following two markers, should they exist:
+    ```
+    [comment]: <> (START COMMAND DOCS)
+    [comment]: <> (END COMMAND DOCS)`
+    ```
+    """
+    _process_commands(
+        app=app, app_name=app_name, repo_root=repo_root, docs_dir=docs_dir
+    )
+    _update_readme(
         app=app, app_name=app_name, repo_root=repo_root, readme_path=readme_path
     )
 
@@ -322,11 +334,11 @@ if __name__ == "__main__":
     from dbtwiz.main import app
 
     script_dir = Path(__file__).parent.resolve()
-    repo_root = script_dir.parent.parent  # Goes up two levels from dbtwiz/utils/
+    repo_root = script_dir.parent.parent  # Goes up two levels from dbtwiz/helpers/
     docs_dir = repo_root / "docs"
     readme_path = repo_root / "README.md"
 
-    update_docs(
+    generate(
         app=app,
         app_name="dbtwiz",
         repo_root=repo_root,
